requirements.txt
yfinance>=0.2.31
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
mlflow>=2.9.0
scikit-learn>=1.3.0
xgboost>=2.0.0
evidently>=0.4.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
pandas>=2.1.0
numpy>=1.25.0
plotly>=5.18.0
ta>=0.11.0
python-dotenv>=1.0.0
httpx>=0.25.0
streamlit-autorefresh>=1.0.0
